✨ Itemized plugin-interface pairs
Resolves #54
