+++ PlugHub.UnitTests/Services/PluginResolverTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging.Abstractions;
 using PlugHub.Services;
 using PlugHub.Shared;
+using PlugHub.Shared.Models;
 
 namespace PlugHub.UnitTests.Services
 {
